refactor: all mappers moved to separate folder
