+++ src/main/java/ru/yandex/practicum/filmorate/mapper/ReviewRatingMapper.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.storage.review;
+package ru.yandex.practicum.filmorate.mapper;
 
 import org.springframework.jdbc.core.RowMapper;
