+++ src/main/java/ru/yandex/practicum/filmorate/storage/event/DbEventStorage.java
@@ -3,6 +3,7 @@ package ru.yandex.practicum.filmorate.storage.event;
 import lombok.RequiredArgsConstructor;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Component;
+import ru.yandex.practicum.filmorate.mapper.EventMapper;
 import ru.yandex.practicum.filmorate.model.Event;
 
 import java.util.List;
