+++ src/main/java/ru/yandex/practicum/filmorate/storage/review/DbReviewStorage.java
@@ -5,6 +5,8 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Component;
+import ru.yandex.practicum.filmorate.mapper.ReviewMapper;
+import ru.yandex.practicum.filmorate.mapper.ReviewRatingMapper;
 import ru.yandex.practicum.filmorate.model.Review;
 
 import java.sql.PreparedStatement;
